fix(ui): show device management banners on home screen instead of settings
Banners (device limit reached, account removed, server notifications)
rendered inside MainStack's shared header, so they appeared on every
MainStack screen with a header, including Settings, instead of Home
as designed in Figma. Move NotificationBannerContainer into the v4.1
Home screen (pinned above the scrolling content, mirroring the
HomeV4_0_x reference pattern) and swap MainStack's header back to the
plain bannerless header so Settings and other MainStack screens no
longer show banners. Services and Wallet tabs are unaffected since
they never rendered the banner container.
